hyprtester: Refactor includes +minor
Minor local improvements:

- Move `Colors`, which are always used with `NLog`, to Log.hpp;

- Remove unused includes, sort some includes (not all; this should
  maybe automated with `clang-format`);

- Apply linter's suggestion(s);

- Replace hacky `#undef`s in hyprtester's main with a single flag macro.
