+++ hyprtester/src/Log.hpp
@@ -3,7 +3,16 @@
 #include <format>
 #include <print>
 
-#include "shared.hpp"
+// Stolen from hyprutils
+namespace Colors {
+    constexpr const char* RED     = "\x1b[31m";
+    constexpr const char* GREEN   = "\x1b[32m";
+    constexpr const char* YELLOW  = "\x1b[33m";
+    constexpr const char* BLUE    = "\x1b[34m";
+    constexpr const char* MAGENTA = "\x1b[35m";
+    constexpr const char* CYAN    = "\x1b[36m";
+    constexpr const char* RESET   = "\x1b[0m";
+};
 
 namespace NLog {
     template <typename... Args>
