+++ hyprtester/src/shared.hpp
@@ -1,22 +1,14 @@
-// Stolen from hyprutils
-
 #pragma once
+#include "Log.hpp"
+#include "tests/shared.hpp"
+
 #include <cmath>
+#include <map>
 #include <string>
 
 // TODO: localize these global variables
 inline std::string HIS       = "";
 inline std::string WLDISPLAY = "";
-
-namespace Colors {
-    constexpr const char* RED     = "\x1b[31m";
-    constexpr const char* GREEN   = "\x1b[32m";
-    constexpr const char* YELLOW  = "\x1b[33m";
-    constexpr const char* BLUE    = "\x1b[34m";
-    constexpr const char* MAGENTA = "\x1b[35m";
-    constexpr const char* CYAN    = "\x1b[36m";
-    constexpr const char* RESET   = "\x1b[0m";
-};
 
 // =================================
 //      TEST CASES DEFINITION
