+++ hyprtester/src/tests/clients/pointer-warp.cpp
@@ -1,4 +1,6 @@
 #include "../../hyprctlCompat.hpp"
+#include "../../Log.hpp"
+#include "../../shared.hpp"
 #include "../shared.hpp"
 #include "tests.hpp"
 #include "build.hpp"
