+++ hyprtester/src/tests/main/hyprctl.cpp
@@ -2,14 +2,9 @@
 #include "../../shared.hpp"
 #include "../../hyprctlCompat.hpp"
 #include <cstdint>
-#include <print>
 #include <string>
-#include <thread>
-#include <chrono>
 #include <hyprutils/os/Process.hpp>
 #include <hyprutils/memory/WeakPtr.hpp>
-#include <csignal>
-#include <cerrno>
 #include "../shared.hpp"
 
 using namespace Hyprutils::OS;
