+++ hyprtester/src/tests/main/moveintoorcreategroup.cpp
@@ -1,13 +1,8 @@
 #include "tests.hpp"
 #include "../../shared.hpp"
 #include "../../hyprctlCompat.hpp"
-#include <print>
-#include <thread>
-#include <chrono>
 #include <hyprutils/os/Process.hpp>
 #include <hyprutils/memory/WeakPtr.hpp>
-#include <csignal>
-#include <cerrno>
 #include "../shared.hpp"
 
 using namespace Hyprutils::OS;
