+++ hyprtester/src/tests/main/workspaces.cpp
@@ -1,14 +1,9 @@
 #include "tests.hpp"
 #include "../../shared.hpp"
 #include "../../hyprctlCompat.hpp"
-#include <print>
-#include <thread>
-#include <chrono>
 #include <hyprutils/os/Process.hpp>
 #include <hyprutils/memory/WeakPtr.hpp>
 #include <hyprutils/utils/ScopeGuard.hpp>
-#include <csignal>
-#include <cerrno>
 #include "../shared.hpp"
 
 using namespace Hyprutils::OS;
